Add Storage class
